more cleanup (wip)
