+++ src/composables/useSkin.ts
@@ -1,6 +1,8 @@
 import { ref, computed, watch } from 'vue';
 import { lsGet, lsSet } from '@/helpers/utils';
 import { useDomain } from '@/composables/useDomain';
+import { useApolloQuery } from '@/composables/useApolloQuery';
+import { SPACE_SKIN_QUERY } from '@/helpers/queries';
 
 const NOT_SET = 'not-set';
 const DARK_MODE = 'dark-mode';
@@ -25,7 +27,10 @@ const _toggleSkin = skin => {
 };
 
 export function useSkin() {
+  const { apolloQuery } = useApolloQuery();
   const { domain } = useDomain();
+
+  const skinName = ref('');
 
   function toggleSkin() {
     const currentSkin = lsGet('skin', NOT_SET);
@@ -36,14 +41,29 @@ export function useSkin() {
     }
   }
 
-  function getSkin (skinName: string) {
-    if (domain && skinName !== 'default') {
-      let skinClass = skinName;
+  const skin = computed(() => {
+    if (domain && skinName.value !== 'default') {
+      let skinClass = skinName.value;
       if (userSkin.value === 'dark-mode')
-        skinClass += ` ${skinName}-dark-mode`;
+        skinClass += ` ${skinName.value}-dark-mode`;
       return skinClass;
     }
     return userSkin.value;
+  });
+
+  async function getSkin() {
+    if (domain) {
+      const spaceObj = await apolloQuery(
+        {
+          query: SPACE_SKIN_QUERY,
+          variables: {
+            id: domain
+          }
+        },
+        'space'
+      );
+      skinName.value = spaceObj?.skin;
+    }
   }
 
   watch(
@@ -58,6 +78,7 @@ export function useSkin() {
   );
 
   return {
+    skin,
     userSkin,
     getSkinIcon,
     toggleSkin,
